Beer/Order/License outline +db +dep on commons

+++ src/main/kotlin/com/tsarpirate/shop/model/PriceModel.kt
@@ -4,4 +4,4 @@ package com.tsarpirate.shop.model
  * @param price the beer price in BGN stotinki
  * @param licenseType
  */
-data class PriceModel(val price: Long, val licenseType: String)
+data class PriceModel(val price: Int, val licenseType: String)
